+++ platform/ui/src/components/studyList/StudyListLoadingText.js
@@ -1,5 +1,6 @@
 import React from 'react';
 import { Icon } from './../../elements/Icon';
+// TODO: useTranslation
 import { withTranslation } from '../../utils/LanguageProvider';
 
 function StudyListLoadingText({ t: translate }) {