+++ platform/viewer/cypress/integration/common/OHIFStandaloneViewer.spec.js
@@ -1,14 +1,14 @@
 describe('OHIFStandaloneViewer', () => {
   beforeEach(() => {
-    cy.visit('/');
+    cy.openStudyList();
   });
 
   it('loads route with at least 2 rows', () => {
+    cy.screenshot();
+    cy.percyCanvasSnapshot('Study List');
+
     cy.get('[data-cy="study-list-results"] tr')
       .its('length')
       .should('be.gt', 2);
-
-    cy.screenshot();
-    cy.percySnapshot('Study List', { width: [340, 768, 1000, 1780] });
   });
 });