+++ platform/viewer/cypress/integration/pwa/OHIFExtensionHTML.spec.js
@@ -1,7 +1,7 @@
 describe('OHIF HTML Extension', () => {
   before(() => {
     cy.openStudy('Dummy');
-    cy.expectMinimumThumbnails(1);
+    cy.expectMinimumThumbnails(5);
   });
 
   it('checks if series thumbnails are being displayed', () => {
@@ -14,6 +14,7 @@ describe('OHIF HTML Extension', () => {
   it('drags and drop a SR thumbnail into viewport', () => {
     cy.get('[data-cy="thumbnail-list"]')
       .contains('SR')
+      .first()
       .drag('.viewport-drop-target');
 
     cy.get(':nth-child(2) > h1').should(