+++ platform/viewer/cypress/integration/pwa/OHIFExtensionMicroscopy.spec.js
@@ -1,26 +1,27 @@
-describe('OHIF PDF Extension', () => {
+describe('OHIF Microscopy Extension', () => {
   before(() => {
-    cy.openStudy('Dummy');
+    cy.openStudyModality('SM');
     cy.expectMinimumThumbnails(6);
   });
 
   it('checks if series thumbnails are being displayed', () => {
     cy.get('[data-cy="thumbnail-list"]')
-      .contains('DOC')
+      .contains('SM')
       .its('length')
       .should('to.be.at.least', 1);
   });
 
-  it('drags and drop a PDF thumbnail into viewport', () => {
+  it('drags and drop a SM thumbnail into viewport', () => {
     cy.get('[data-cy="thumbnail-list"]')
-      .contains('DOC')
+      .contains('SM')
       .drag('.viewport-drop-target');
 
-    cy.get('.DicomPDFViewport')
+    cy.get('.DicomMicroscopyViewer')
       .its('length')
       .should('be.eq', 1);
 
+    cy.wait(3000); //Waiting for image to render before taking the snapshot
     cy.screenshot();
-    cy.percyCanvasSnapshot('PDF Extension');
+    cy.percyCanvasSnapshot('Microscopy Extension');
   });
 });